+++ dqops/src/main/java/com/dqops/core/configuration/DqoUiConfigurationProperties.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright © 2021 DQOps (support@dqops.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.dqops.core.configuration;
+
+import lombok.EqualsAndHashCode;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration POJO with the configuration for the dqo.ui that some activities how the UI works.
+ */
+@Configuration
+@ConfigurationProperties(prefix = "dqo.ui")
+@EqualsAndHashCode(callSuper = false)
+public class DqoUiConfigurationProperties implements Cloneable {
+    private String applicationName = "DQOps Data Quality Operations Center";
+
+    /**
+     * Returns the application name.
+     * @return Application name.
+     */
+    public String getApplicationName() {
+        return applicationName;
+    }
+
+    /**
+     * Sets the application name.
+     * @param applicationName Application name.
+     */
+    public void setApplicationName(String applicationName) {
+        this.applicationName = applicationName;
+    }
+
+    /**
+     * Creates a clone of the object.
+     * @return Cloned instance.
+     */
+    @Override
+    public DqoUiConfigurationProperties clone() {
+        try {
+            DqoUiConfigurationProperties cloned = (DqoUiConfigurationProperties) super.clone();
+            return cloned;
+        }
+        catch (CloneNotSupportedException ex) {
+            throw new RuntimeException("Cannot clone object", ex);
+        }
+    }
+}