+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 1.11.3
 
-* 
+* Fixes to use parameters for custom checks.