+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 1.11.3
 
 * Fixes to use parameters for custom checks.
+* Shorten the time in the connection pool to 10 sec